mango_requirements.txt
####part 1
about-time==4.2.1
absl-py==2.1.0
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.13
alive-progress==3.1.5
antlr4-python3-runtime==4.8
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
babel==2.14.0
backcall==0.2.0
backports-zoneinfo==0.2.1
beautifulsoup4==4.13.3
biopython==1.79
bleach==6.1.0
blinker==1.7.0
botocore==1.34.51
botorch==0.8.5
cachetools==5.0.0
cffi==1.16.0
chemprop==1.6.1
click==8.1.7
cma==2.7.0
cmake==3.31.4
comm==0.2.2
conda-pack==0.7.1
configargparse==1.5.2
cython==3.0.0a10
debugpy==1.8.12
decorator==5.2.1
deepchem==2.6.0.dev20211207173323
defusedxml==0.7.1
deprecated==1.2.14
descriptastorus==2.6.1
dill==0.3.7
django==4.2.9
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-parser==0.15
docutils==0.20.1
evoxbench==1.0.3
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.13.1
fire==0.6.0
flask==3.0.0
flatbuffers==2.0.7
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
gast==0.4.0
gitdb==4.0.12
gitpython==3.1.44
glfw==2.7.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpytorch==1.10
grapheme==0.6.0
grpcio==1.70.0
gym==0.14.0
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
higher==0.2.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.0.8
hydra-core==1.1.0
hyperopt==0.2.7
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
jinja2==3.1.3
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keras==2.7.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
lightning-utilities==0.11.9
linear-operator==0.4.0
lit==18.1.8
markdown==3.7
markupsafe==2.1.3
matplotlib-inline==0.1.7
mistune==3.1.2
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook-shim==0.2.4
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
####part 2
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
panda==0.3.1
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pip==24.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
promise==2.3
prompt-toolkit==3.0.50
propcache==0.2.0
protobuf==3.20.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.21
pydeprecate==0.3.2
pyglet==1.3.2
pygments==2.17.2
pymoo==0.5.0
pyro-api==0.1.2
pyro-ppl==1.8.6
python-json-logger==3.2.1
pyyaml==6.0.1
pyzmq==26.2.1
randomname==0.2.1
rdkit==2023.9.4
rdkit-pypi==2022.9.5
referencing==0.35.1
regex==2023.12.25
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9
s3fs==2024.3.1
sacremoses==0.1.1
scikit-learn==0.21.3
scipy==1.10.1
selfies==2.0.0
send2trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.5
shortuuid==1.0.13
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.8.9
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.13.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.10.3
tomli==2.2.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.6.0
triton==2.0.0
typed-argument-parser==1.9.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
uri-template==1.3.0
urllib3==1.26.18
wandb==0.12.21
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2023.1.0
yarl==1.9.4
zipp==3.17.0


-i https://pypi.tuna.tsinghua.edu.cn/simple